{{cookiecutter.directory}}/pyproject.toml
[project]
name = "{{cookiecutter.mypackage}}"
version = "0.1.0"
description = "Python Package for the paper: {{cookiecutter.paper}}"
authors = [
    {name = "Your Name", email = "abc@gmail.com"}
    ]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "forallpeople>=2.7.0",
    "holoviews>=1.19.1",
    "hvplot>=0.11.0",
    "icecream>=2.1.3",
    "invoke>=2.2.0",
    "jupyterlab>=4.2.5",
    "jupytext>=1.16.4",
    "matplotlib>=3.9.2",
    "notebook>=7.2.2",
    "numpy>=2.1.2",
    "pandas>=2.2.3",
    "pyprojroot>=0.3.0",
    "pytest>=8.4.2",
    "python-box>=7.2.0",
    "seaborn>=0.13.2",
    "si-prefix>=1.3.3",
    "sigfig>=1.3.3",
    "sitepath>=0.4.0",
    "summarytools>=0.3.0",
    "typer>=0.16.1",
    "watermark>=2.5.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.jupytext]
formats = "ipynb,py:percent"
notebook_metadata_filter="-all"

[project.scripts]
app = "{{cookiecutter.mypackage}}.scripts:app"
